src/modules/checkout/facade/checkout.facade.ts
import UseCaseInterface from "../../@shared/usecase/usecase.interface";
import CheckoutFacadeInterface, { PlaceOrderFacadeInputDto, PlaceOrderFacadeOutputDto } from "./checkout.facade.interface";

export interface UseCasesProps {
    placeOrderUseCase: UseCaseInterface;
}

export default class CheckoutFacade implements CheckoutFacadeInterface {
    private _placeOrderUseCase: UseCaseInterface;

    constructor(usecaseProps: UseCasesProps) {
        this._placeOrderUseCase = usecaseProps.placeOrderUseCase;
    }

    async placeOrder(input: PlaceOrderFacadeInputDto): Promise<PlaceOrderFacadeOutputDto> {
        return await this._placeOrderUseCase.execute(input);
    }
}